Updated Polynomial Class design
I added default values for the parameters of the initializer. I opted for the signs of the coefficient to be randomly assigned instead of relying on the user to include negative numbers in their domain; the domain has to be greater than zero (n < 0). 
Finally, I added documentation for the initializer.